+++ src/geocat/comp/__init__.py
@@ -14,7 +14,7 @@ from .meteorology import (dewtemp, heat_index, showalter_index, relhum,
                           relhum_ice, relhum_water,
                           actual_saturation_vapor_pressure, max_daylight,
                           psychrometric_constant, saturation_vapor_pressure,
-                          saturation_vapor_pressure_slope)
+                          saturation_vapor_pressure_slope, calc_deltapressure)
 from .skewt_params import get_skewt_vars, showalter_index
 from .spherical import decomposition, recomposition, scale_voronoi
 from .stats import eofunc, eofunc_eofs, eofunc_pcs, eofunc_ts, pearson_r
